+++ src/main/kotlin/dev/turingcomplete/intellijjvmsmanagerplugin/process/ProcessType.kt
@@ -11,8 +11,6 @@ enum class ProcessType(val description: String?, private val loadIcon: (() -> Ic
   // -- Values ------------------------------------------------------------------------------------------------------ //
 
   INTELLIJ_IDEA("IntelliJ IDEA", "intellij.svg"),
-  INTELLIJ_IDEA_FSNOTIFIER("IntelliJ IDEA File Watcher", "intellij.svg", "https://github.com/int128/idea-fsnotifier-wsl"),
-  INTELLIJ_IDEA_COMPILER_PROCESS("IntelliJ IDEA Compiler Process", "intellij.svg"),
   GRADLE_DAEMON("Gradle Daemon", "gradle.svg", "https://docs.gradle.org/current/userguide/gradle_daemon.html"),
   GRADLE_WORKER("Gradle Worker", "gradle.svg", "https://docs.gradle.org/current/userguide/worker_api.html"),
   GRADLE_WRAPPER("Gradle Wrapper", "gradle.svg", "https://docs.gradle.org/current/userguide/gradle_wrapper.html"),
@@ -27,18 +25,19 @@ enum class ProcessType(val description: String?, private val loadIcon: (() -> Ic
   // -- Companion Object -------------------------------------------------------------------------------------------- //
 
   companion object {
-    private val processNameToType = mapOf(Pair("javac", JAVAC), Pair("fsnotifier", INTELLIJ_IDEA_FSNOTIFIER), Pair("jcef", JCEF),
+    private val processNameToType = mapOf(Pair("javac", JAVAC), Pair("fsnotifier", INTELLIJ_IDEA), Pair("jcef", JCEF),
                                           Pair("git", GIT), Pair("idea", INTELLIJ_IDEA), Pair("idea64", INTELLIJ_IDEA))
 
     fun determine(process: OSProcess, vmDescriptor: VirtualMachineDescriptor?): ProcessType {
+      val vmDisplayName = vmDescriptor?.displayName()
       return processNameToType[process.name] ?: when {
-        process.commandLine?.contains("com.intellij.idea.Main") == true -> INTELLIJ_IDEA
-        process.commandLine?.contains("org.jetbrains.jps.cmdline.Launcher") == true -> INTELLIJ_IDEA_COMPILER_PROCESS
-        process.commandLine?.contains("org.gradle.launcher.daemon.bootstrap.GradleDaemon") == true -> GRADLE_DAEMON
-        process.commandLine?.contains("worker.org.gradle.process.internal.worker.GradleWorkerMain") == true -> GRADLE_WORKER
-        process.commandLine?.contains("org.gradle.wrapper.GradleWrapperMain") == true -> GRADLE_WRAPPER
-        process.commandLine?.contains("org.jetbrains.kotlin.daemon.KotlinCompileDaemon") == true -> KOTLIN_COMPILE_DAEMON
-        process.commandLine?.contains("org.apache.maven.wrapper.MavenWrapperMain") == true -> MAVEN_WRAPPER
+        vmDisplayName?.startsWith("com.intellij") == true -> INTELLIJ_IDEA
+        vmDisplayName?.startsWith("org.jetbrains.kotlin.daemon.KotlinCompileDaemon") == true -> KOTLIN_COMPILE_DAEMON
+        vmDisplayName?.startsWith("org.jetbrains") == true -> INTELLIJ_IDEA
+        vmDisplayName?.startsWith("org.gradle.launcher.daemon.bootstrap.GradleDaemon") == true -> GRADLE_DAEMON
+        vmDisplayName?.startsWith("worker.org.gradle.process.internal.worker.GradleWorkerMain") == true -> GRADLE_WORKER
+        vmDisplayName?.startsWith("org.gradle.wrapper.GradleWrapperMain") == true -> GRADLE_WRAPPER
+        vmDisplayName?.startsWith("org.apache.maven.wrapper.MavenWrapperMain") == true -> MAVEN_WRAPPER
         process.name == "java" || vmDescriptor != null -> JAVA
         else -> UNKNOWN
       }